Refactoring

+++ src/main/java/spring/study/board/service/BoardService.java
@@ -103,7 +103,7 @@ public class BoardService {
     }
 
     @Transactional
-    public int updateBoard(Long id, String content) {
+    public long updateBoard(Long id, String content) {
         Board board = boardRepository.findById(id).orElseThrow(() -> new RuntimeException(
                 "존재하지 않는 게시글입니다."
         ));
@@ -111,7 +111,7 @@ public class BoardService {
         board.changeContent(content);
         board.changeUpdateTime(LocalDateTime.now());
 
-        return board.getId().intValue();
+        return board.getId();
     }
 
     public void deleteById(Long id) {
